Updated Sage Line properties query to remove mv table
The query that looked up all the line properties fromSAGE was using a
materialized view. This caused the pipeline to fail when the view was
not updated for a number of days. I also believe that it was the
underlying cause of a lot of problems and hangups during the sage
discovery step as the updating of the materialized view was on a 15
minute timer, which means it could have been out of date for up to 15
minutes. So, when TMOG inserts data into SAGE and then tells JACS to go
get it and JACS looks in the materialized view, it finds nothing.
ERROR!!!


Former-commit-id: 8863852b429c7729c6c68dd4d296bfe91bb02d1b